refactor test and code structure for text processing
