+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 # See https://pre-commit.com for more information
 # See https://pre-commit.com/hooks.html for more hooks
 exclude: '^$'
-fail_fast: false
+fail_fast: true
 repos:
 -   repo: local
     hooks:
