+++ tests/unit/process_text/test_extract_all_skills_from_text.py
@@ -1,4 +1,4 @@
-from src.ocr import extract_all_skills_from_text
+from src.process_text import extract_all_skills_from_text
 
 
 def test_extract_all_skills_from_text(prep_extract_all_skills_from_text):
